Fix authorization errors

+++ ApiContestNew/Controllers/AnimalTypeController.cs
@@ -40,6 +40,11 @@ namespace ApiContestNew.Controllers
         [HttpPost]
         public async Task<ActionResult<GetAnimalTypeDto>> AddAnimalType(AddAnimalTypeDto dto)
         {
+            if (string.IsNullOrWhiteSpace(Request.Headers.Authorization))
+            {
+                return Unauthorized();
+            }
+
             var response = await _animalTypeService.AddAnimalTypeAsync(_mapper.Map<AnimalType>(dto));
 
             return response.StatusCode switch
@@ -54,6 +59,11 @@ namespace ApiContestNew.Controllers
         [HttpPut("{typeId}")]
         public async Task<ActionResult<GetAnimalTypeDto>> UpdateAnimalType(long typeId, UpdateAnimalTypeDto dto)
         {
+            if (string.IsNullOrWhiteSpace(Request.Headers.Authorization))
+            {
+                return Unauthorized();
+            }
+
             var response = await _animalTypeService.UpdateAnimalTypeAsync(typeId, _mapper.Map<AnimalType>(dto));
 
             return response.StatusCode switch
@@ -69,6 +79,11 @@ namespace ApiContestNew.Controllers
         [HttpDelete("{typeId}")]
         public async Task<ActionResult<GetAnimalTypeDto>> DeleteAnimalType(long typeId)
         {
+            if (string.IsNullOrWhiteSpace(Request.Headers.Authorization))
+            {
+                return Unauthorized();
+            }
+
             var response = await _animalTypeService.DeleteAnimalTypeAsync(typeId);
 
             return response.StatusCode switch
